pyproject.toml
[tool.poetry]
name = "translate_linebot"
version = "0.1.0"
description = ""
authors = ["Masayuki1009 <smasayukii1009@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.79.0"
uvicorn = {extras = ["standard"], version = "^0.18.2"}
line-bot-sdk = "^2.3.0"
deepl = "^1.9.0"
python-dotenv = "^0.20.0"
SQLAlchemy = "^1.4.39"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.6.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
